+++ src/Job/Job.php
@@ -51,21 +51,21 @@ class Job implements JobInterface
      * @param string|DateTimeInterface $startDate - @see DateTime supported formats
      * @param callable $callback
      * @param string|DateTimeZone $timezone - If $timezone is omitted, the current timezone will be used.
+     *                                        If $startDate is instance of `DateTimeInterface` then $timezone parameter will be ignored
      * @return Job
      */
     public static function createFromString($rRule, $startDate, callable $callback, $timezone = null)
     {
-        if (empty($timezone)) {
+        if ($timezone === null) {
             $timezone = date_default_timezone_get();
         }
         if (is_string($timezone)) {
-            $timezone = new \DateTimeZone($timezone);
+            $timezone = new DateTimeZone($timezone);
         }
         if (!$startDate instanceof DateTimeInterface) {
             $startDate = new DateTime($startDate, $timezone);
         }
-        $timezone = $startDate->getTimezone()->getName();
-        $rRule = new RRule($rRule, $startDate, $timezone);
+        $rRule = new RRule($rRule, $startDate);
         return new self($rRule, $callback);
     }
 }