+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.2.1 - 2018-01-03
+
+### Fixed
+- Fix timezone processing in `Job::createFromString()`
+
 ## 0.2.0 - 2018-01-03
 
 ### Changed